Add isFOO_utf8_safe() macros
The original API assumed that we could keep malformed UTF-8 out by use
of gatekeepers, but that is currently impossible.  This commit adds
"safe" versions to macros for determining if a UTF-8 sequence represents
an alphabetic, a digit, etc.  Each new macro has an extra parameter
pointing to the end of the sequence, so that looking beyond the input
string can be avoided.

The macros aren't currently completely safe, as they don't test that
there is at least a single valid character in the input, except by an
assertion in DEBUGGING builds.  This is because typically they are
called in code that makes that assumption, and frequently tests the
current character for one thing or another.  While debugging this and
future commits, The assertion showed some current errors where that
assumption turned out to be false.