PATCH: [perl #129953] lib/locale.t failures on FREEBSD
Because collation is very complicated, perl has opted to use the one
furnished in libc.  And because Perl strings are different from C
strings in that the former allows embedded NULs, something needs to be
done when collating strings that contain NULs.  What happens is that
when calculating the collation, each embedded NUL is replaced by the
lowest collating control character in the range 1-255.  If no suitable
control character is found, the lowest collating code point is used.
(The reason controls are preferred is that collation is context
sensitive, with adjacent characters potentially affecting the collation
order, and controls are much less likely to have any such interactions.)

This bug occurred if the character replacing the NUL is UTF-8 variant.
Basically it got doubly encoded as UTF-8.  On most systems the NUL
replacement is \x01, which is invariant under UTF-8, but in freebsd, the
replacement is \x80, which is UTF-8 variant.   The solution is to store
the replacement as a code point, and encode it as UTF-8 only if
necessary, once.  This actually simplifies the code